+++ totalRP3/Core/Globals.lua
@@ -35,7 +35,7 @@ TRP3_API.globals = {
 	addon_name_me = "Total RP 3",
 	addon_id_length = 15,
 
-	version = 109,
+	version = 111,
 
 	--@debug@
 	version_display = "-dev",