+++ totalRP3/Locales/Locale.lua
@@ -1423,12 +1423,6 @@ If you wish to report %s's profile and you cannot target them you will need to o
 - Removed temporarily right-click menu options in Retail due to a conflict with Edit Mode that can prevent the use of abilities in combat. They will be re-added once the conflict has been resolved.
 
 ]],
-
-	------------------------------------------------------------------------------------------------
-	--- PLACE LOCALIZATION NOT ALREADY UPLOADED TO CURSEFORGE HERE
-	--- THEN MOVE IT UP ONCE IMPORTED
-	------------------------------------------------------------------------------------------------
-
 	WHATS_NEW_25_2 = [[# Changelog version 2.4.1
 
 ## Added
@@ -1442,6 +1436,54 @@ If you wish to report %s's profile and you cannot target them you will need to o
 - Fixed the "Hide immediately instead of fading" tooltip setting.
 - Fixed another issue with water elemental pets not being valid targets to bind a companion profile to in WotLK Classic
 - Fixed an issue with colorblind mode preventing companion tooltips from displaying.
+
+]],
+	WHATS_NEW_25_3 = [[# Changelog version 2.4.2
+
+## Fixed
+
+- Fixed missing dracthyr default profile icon.
+- Fixed companion profiles not working properly on the second pet from Animal Companion talent.
+- Fixed companion profiles on water elementals in WotLK Classic (for real this time).
+- Fixed broadcast channel order setting being checked before configuration was set up.
+- Fixed an issue with chat links that could lead to blocked actions.
+
+]],
+
+	------------------------------------------------------------------------------------------------
+	--- PLACE LOCALIZATION NOT ALREADY UPLOADED TO CURSEFORGE HERE
+	--- THEN MOVE IT UP ONCE IMPORTED
+	------------------------------------------------------------------------------------------------
+
+	CO_DATE_FORMAT = "Date format",
+	CO_DATE_FORMAT_HELP = "Format string used for date/time strings in the addon.|n|nIf left empty, this will display date/time strings in a format appropriate to the configured locale of the addon.|n|nThe following format specifiers are supported:|n|n%1$s|n|nAll other characters are treated literally.",
+	CO_DATE_FORMAT_SPEC_a = "%1$s: Abbreviated weekday name (%2$s)",
+	CO_DATE_FORMAT_SPEC_A = "%1$s: Full weekday name (%2$s)",
+	CO_DATE_FORMAT_SPEC_b = "%1$s: Abbreviated month name (%2$s)",
+	CO_DATE_FORMAT_SPEC_B = "%1$s: Full month name (%2$s)",
+	CO_DATE_FORMAT_SPEC_d = "%1$s: Day of the month (%2$s)",
+	CO_DATE_FORMAT_SPEC_H = "%1$s: Hour using a 24-hour clock (%2$s)",
+	CO_DATE_FORMAT_SPEC_I = "%1$s: Hour using a 12-hour clock (%2$s)",
+	CO_DATE_FORMAT_SPEC_m = "%1$s: Month (%2$s)",
+	CO_DATE_FORMAT_SPEC_M = "%1$s: Minute (%2$s)",
+	CO_DATE_FORMAT_SPEC_p = "%1$s: Either 'am' or 'pm' (%2$s)",
+	CO_DATE_FORMAT_SPEC_S = "%1$s: Second (%2$s)",
+	CO_DATE_FORMAT_SPEC_y = "%1$s: Two-digit year (%2$s)",
+	CO_DATE_FORMAT_SPEC_Y = "%1$s: Full year (%2$s)",
+	CO_DATE_FORMAT_SPEC_ESC = "%1$s: The literal character '%%'",
+	WHATS_NEW_25_4 = [[# Changelog version 2.4.3
+
+## Added
+
+- Added a setting to customize date formats displayed in the directory. (Thanks to Kisis for the initial work on it)
+  - The default format now respects regional formats based on current game client.
+  - Total RP 3: Extended will be updated to support this at a later date.
+
+## Fixed
+
+- Fixed tooltipData error showing since 10.0.2.
+- Fixed map scan not working on WotLK Classic.
+- Fixed potential taint issues leading to blocked actions.
 
 ]],