+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# Changelog version 2.4.3
+
+## Added
+
+- Added a setting to customize date formats displayed in the directory. (Thanks to Kisis for the initial work on it)
+  - The default format now respects regional formats based on current game client.
+  - Total RP 3: Extended will be updated to support this at a later date.
+
+## Fixed
+
+- Fixed tooltipData error showing since 10.0.2.
+- Fixed map scan not working on WotLK Classic.
+- Fixed potential taint issues leading to blocked actions.
+
+# Changelog version 2.4.2
+
+## Fixed
+
+- Fixed missing dracthyr default profile icon.
+- Fixed companion profiles not working properly on the second pet from Animal Companion talent.
+- Fixed companion profiles on water elementals in WotLK Classic (for real this time).
+- Fixed broadcast channel order setting being checked before configuration was set up.
+- Fixed an issue with chat links that could lead to blocked actions.
+
 # Changelog version 2.4.1
 
 ## Added